testing/web-platform/tests/html/semantics/interfaces.html

<!DOCTYPE html>
<title>Test of interfaces</title>
<link rel="author" title="Ms2ger" href="mailto:Ms2ger@gmail.com">
<link rel="help" href="http://www.whatwg.org/html5/">
<link rel="help" href="http://dev.w3.org/2006/webapi/WebIDL/#host-objects">
<link rel="help" href="http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-262.pdf#page=96">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
var data = [
  ["abbr", ""],
  ["acronym", ""],
  ["address", ""],
  ["applet", "Applet"],
  ["audio", "Audio"],
  ["b", ""],
  ["base", "Base"],
  ["basefont", "Unknown"],
  ["bdo", ""],
  ["bgsound", "Unknown"],
  ["big", ""],
  ["blink", "Unknown"],
  ["blockquote", "Quote"],
  ["body", "Body"],
  ["br", "BR"],
  ["button", "Button"],
  ["canvas", "Canvas"],
  ["caption", "TableCaption"],
  ["center", ""],
  ["cite", ""],
  ["code", ""],
  ["col", "TableCol"],
  ["colgroup", "TableCol"],
  ["dd", ""],
  ["del", "Mod"],
  ["dfn", ""],
  ["dir", "Directory"],
  ["div", "Div"],
  ["dl", "DList"],
  ["dt", ""],
  ["em", ""],
  ["embed", "Embed"],
  ["fieldset", "FieldSet"],
  ["font", "Font"],
  ["form", "Form"],
  ["frame", "Frame"],
  ["frameset", "FrameSet"],
  ["h1", "Heading"],
  ["h2", "Heading"],
  ["h3", "Heading"],
  ["h4", "Heading"],
  ["h5", "Heading"],
  ["h6", "Heading"],
  ["head", "Head"],
  ["hr", "HR"],
  ["html", "Html"],
  ["i", ""],
  ["iframe", "IFrame"],
  ["image", "Unknown"],
  ["img", "Image"],
  ["input", "Input"],
  ["ins", "Mod"],
  ["isindex", "Unknown"],
  ["kbd", ""],
  ["keygen", "Keygen"],
  ["label", "Label"],
  ["legend", "Legend"],
  ["li", "LI"],
  ["link", "Link"],
  ["listing", ""],
  ["map", "Map"],
  ["marquee", "Marquee"],
  ["menu", "Menu"],
  ["meta", "Meta"],
  ["multicol", "Unknown"],
  ["nobr", ""],
  ["noembed", ""],
  ["noframes", ""],
  ["noscript", ""],
  ["object", "Object"],
  ["ol", "OList"],
  ["optgroup", "OptGroup"],
  ["option", "Option"],
  ["p", "Paragraph"],
  ["param", "Param"],
  ["plaintext", ""],
  ["pre", "Pre"],
  ["q", "Quote"],
  ["s", ""],
  ["samp", ""],
  ["script", "Script"],
  ["select", "Select"],
  ["small", ""],
  ["source", "Source"],
  ["spacer", "Unknown"],
  ["span", "Span"],
  ["strike", ""],
  ["strong", ""],
  ["style", "Style"],
  ["sub", ""],
  ["sup", ""],
  ["table", "Table"],
  ["tbody", "TableSection"],
  ["td", "TableDataCell"],
  ["textarea", "TextArea"],
  ["tfoot", "TableSection"],
  ["th", "TableHeaderCell"],
  ["thead", "TableSection"],
  ["title", "Title"],
  ["tr", "TableRow"],
  ["tt", ""],
  ["u", ""],
  ["ul", "UList"],
  ["var", ""],
  ["video", "Video"],
  ["wbr", ""],
  ["xmp", ""],
  ["section", ""],
  ["nav", ""],
  ["article", ""],
  ["aside", ""],
  ["hgroup", ""],
  ["header", ""],
  ["footer", ""],
  ["time", "Time"],
  ["mark", ""],
  ["progress", "Progress"],
  ["meter", "Meter"],
  ["ruby", ""],
  ["rt", ""],
  ["rp", ""],
  ["figure", ""],
  ["datalist", "DataList"],
  ["output", "Output"],
  ["details", "Details"],
  ["command", "Command"],
  ["nextid", "Unknown"],
  ["rb", "Unknown"],
  ["bdi", ""],
  ["data", "Data"],
  ["a", "Anchor"],
  ["area", "Area"],
  ["dialog", "Dialog"],
  ["figcaption", ""],
  ["summary", ""],
  ["track", "Track"]
];
</script>
<div id="log"></div>
<script>
test(function() {
  data.forEach(function(a) {
    test(function() {
      var e = document.createElement(a[0]), i = "HTML" + a[1] + "Element";
      assert_class_string(e, i,
                          "Element " + a[0] + " should have " + i +
                          " as its primary interface.");
      assert_true(e instanceof window[i],
                  "Element " + a[0] + " should implement " + i + ".");
      assert_true(e instanceof HTMLElement,
                  "Element " + a[0] + " should implement HTMLElement.");
      assert_true(e instanceof Element,
                  "Element " + a[0] + " should implement Element.");
      assert_true(e instanceof Node,
                  "Element " + a[0] + " should implement Node.");
    }, "Interfaces for " + a[0])
  })
})
</script>
